+++ NEWS.md
@@ -6,8 +6,8 @@
 
 ## Updates of Existing Functions
 
-- `is_valid_dtc()` was enhanced with arguments `valid_full_date` and `valid_full_datetime` to allow users
-to check whether the date/datetime is valid, i.e. parseable by `ymd()` or `ymd_hms()`. (#505)
+- `is_valid_dtc()` was enhanced with the arguments `check_dtc` to allow users
+to check whether the input date/datetime is a real date/datetime`. (#505)
 - `capture_message()` was updated to strip off all control sequences from the
  message. (#543)
 - `assert_list_of()` (and `assert_vars()`) was updated with a clearer message to users on how to fix